+++ NWN.Anvil.Tests/src/main/API/TwoDimArray/TwoDimArrayTests.cs
@@ -30,7 +30,7 @@ namespace Anvil.Tests.API
       ResourceManager.WriteTempResource(resourceName, StringHelper.Cp1252Encoding.GetBytes(twoDimArray));
       createdTempResources.Add(resourceName);
 
-      TwoDimArray array = new TwoDimArray(resourceName);
+      TwoDimArray array = NwGameTables.GetTable(resourceName);
 
       Assert.That(array.RowCount, Is.EqualTo(3));
       Assert.That(array.ColumnCount, Is.EqualTo(4));
@@ -52,15 +52,15 @@ namespace Anvil.Tests.API
     {
       Assert.That(() =>
       {
-        TwoDimArray _ = new TwoDimArray("invalidtest");
+        TwoDimArray _ = NwGameTables.GetTable("invalidtest");
       }, Throws.Exception.TypeOf<ArgumentException>());
     }
 
     [Test(Description = "2da arrays are equal if they reference the same pointer.")]
     public void Same2daIsConsideredEqual()
     {
-      TwoDimArray array1 = new TwoDimArray("appearance.2da");
-      TwoDimArray array2 = new TwoDimArray("appearance.2da");
+      TwoDimArray array1 = NwGameTables.GetTable("appearance.2da");
+      TwoDimArray array2 = NwGameTables.GetTable("appearance.2da");
 
       Assert.That(array1, Is.EqualTo(array2));
     }
@@ -68,8 +68,8 @@ namespace Anvil.Tests.API
     [Test(Description = "2da arrays are equal if they reference the same pointer.")]
     public void Same2daGenericIsConsideredEqual()
     {
-      TwoDimArray<AppearanceTableEntry> array1 = new TwoDimArray<AppearanceTableEntry>("appearance.2da");
-      TwoDimArray<AppearanceTableEntry> array2 = new TwoDimArray<AppearanceTableEntry>("appearance.2da");
+      TwoDimArray<AppearanceTableEntry> array1 = NwGameTables.GetTable<AppearanceTableEntry>("appearance.2da");
+      TwoDimArray<AppearanceTableEntry> array2 = NwGameTables.GetTable<AppearanceTableEntry>("appearance.2da");
 
       Assert.That(array1, Is.EqualTo(array2));
     }
@@ -77,8 +77,8 @@ namespace Anvil.Tests.API
     [Test(Description = "2da arrays are equal if they reference the same pointer.")]
     public void Same2daMixedIsConsideredEqual()
     {
-      TwoDimArray array1 = new TwoDimArray("appearance.2da");
-      TwoDimArray<AppearanceTableEntry> array2 = new TwoDimArray<AppearanceTableEntry>("appearance.2da");
+      TwoDimArray array1 = NwGameTables.GetTable("appearance.2da");
+      TwoDimArray<AppearanceTableEntry> array2 = NwGameTables.GetTable<AppearanceTableEntry>("appearance.2da");
 
       Assert.That(array1, Is.EqualTo(array2));
     }
@@ -86,8 +86,8 @@ namespace Anvil.Tests.API
     [Test(Description = "2da arrays are not equal if they do not reference the same pointer.")]
     public void Different2daAreNotConsideredEqual()
     {
-      TwoDimArray array1 = new TwoDimArray<AppearanceTableEntry>("appearance.2da");
-      TwoDimArray array2 = new TwoDimArray<AppearanceTableEntry>("environment.2da");
+      TwoDimArray array1 = NwGameTables.GetTable<AppearanceTableEntry>("appearance.2da");
+      TwoDimArray array2 = NwGameTables.GetTable<EnvironmentPreset>("environment.2da");
 
       Assert.That(array1, Is.Not.EqualTo(array2));
     }