+++ NWN.Anvil/src/main/API/Object/NwItem.cs
@@ -184,6 +184,7 @@ namespace Anvil.API
     /// <summary>
     /// Gets the GameObject that has this item in its inventory. Returns null if it is on the ground, or not in any inventory.
     /// </summary>
+    /// <remarks>This can be a creature, a placeable, or an item container (e.g magic bag). Use <see cref="RootPossessor"/> to get the root possessor of this item.</remarks>
     public NwGameObject? Possessor => NWScript.GetItemPossessor(this).ToNwObject<NwGameObject>();
 
     /// <summary>
@@ -215,6 +216,24 @@ namespace Anvil.API
     }
 
     /// <summary>
+    /// Gets the root possessor of this item.
+    /// </summary>
+    /// <remarks>If this item is in a container, this is the creature/placeable holding the container that holds this item. Otherwise, this returns the same object as <see cref="Possessor"/>.</remarks>
+    public NwGameObject? RootPossessor
+    {
+      get
+      {
+        NwGameObject? possessor = Possessor;
+        if (possessor is NwItem item)
+        {
+          return item.Possessor as NwCreature;
+        }
+
+        return possessor as NwCreature;
+      }
+    }
+
+    /// <summary>
     /// Gets or sets the weight of this item, in pounds.
     /// </summary>
     public decimal Weight
@@ -223,7 +242,12 @@ namespace Anvil.API
       set
       {
         Item.m_nWeight = (int)Math.Round(value * 10.0m, MidpointRounding.ToZero);
-        Item.m_oidPossessor.ToNwObject<NwCreature>()?.Creature.UpdateEncumbranceState();
+        if (RootPossessor is NwCreature creature)
+        {
+          creature.Creature.m_nEquippedWeight = creature.Creature.ComputeTotalEquippedWeight();
+          creature.Creature.m_nTotalWeightCarried = creature.Creature.ComputeTotalWeightCarried();
+          creature.Creature.UpdateEncumbranceState();
+        }
       }
     }