+++ NWN.Anvil/src/main/Native/CExoStringData.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Runtime.InteropServices;
+using NWN.Native.API;
 
 namespace Anvil.Native
 {
@@ -8,5 +9,10 @@ namespace Anvil.Native
   {
     public readonly IntPtr m_sString; // char*
     public readonly uint m_nBufferLength; // uint32_t
+
+    public override string? ToString()
+    {
+      return m_sString != IntPtr.Zero ? m_sString.ReadNullTerminatedString() : null;
+    }
   }
 }